Refactors typing files and reduces `Any` usage
